ai-rag-service/requirements.txt
fastapi
uvicorn
langchain
faiss-cpu
openai
pydantic
python-dotenv
numpy
